HIVE-27598 - Enhance alter table compact to work for partitioned tables without specifying any partition
