+++ ql/src/java/org/apache/hadoop/hive/ql/txn/compactor/Initiator.java
@@ -86,11 +86,9 @@ import static org.apache.hadoop.hive.conf.Constants.COMPACTOR_INTIATOR_THREAD_NA
  * A class to initiate compactions.  This will run in a separate thread.
  * It's critical that there exactly 1 of these in a given warehouse.
  */
-public class Initiator extends MetaStoreCompactorThread {
+public class Initiator extends InitiatorBase {
   static final private String CLASS_NAME = Initiator.class.getName();
   static final private Logger LOG = LoggerFactory.getLogger(CLASS_NAME);
-
-  static final private String COMPACTORTHRESHOLD_PREFIX = "compactorthreshold.";
 
   private ExecutorService compactionExecutor;
 
@@ -250,36 +248,6 @@ public class Initiator extends MetaStoreCompactorThread {
             MetastoreConf.ConfVars.COMPACTOR_INITIATOR_TABLECACHE_ON);
   }
 
-  protected void scheduleCompactionIfRequired(CompactionInfo ci, Table t, Partition p, String poolName,
-                                            String runAs, boolean metricsEnabled)
-      throws MetaException {
-    StorageDescriptor sd = resolveStorageDescriptor(t, p);
-    try {
-      ValidWriteIdList validWriteIds = resolveValidWriteIds(t);
-
-      checkInterrupt();
-
-      CompactionType type = checkForCompaction(ci, validWriteIds, sd, t.getParameters(), runAs);
-      if (type != null) {
-        ci.type = type;
-        ci.poolName = poolName;
-        requestCompaction(ci, runAs);
-      }
-    } catch (InterruptedException e) {
-      //Handle InterruptedException separately so the compactioninfo won't be marked as failed.
-      LOG.info("Initiator pool is being shut down, task received interruption.");
-    } catch (Throwable ex) {
-      String errorMessage = "Caught exception while trying to determine if we should compact " + ci + ". Marking "
-          + "failed to avoid repeated failures, " + ex;
-      LOG.error(errorMessage);
-      ci.errorMessage = errorMessage;
-      if (metricsEnabled) {
-        Metrics.getOrCreateCounter(MetricsConstants.COMPACTION_INITIATOR_FAILURE_COUNTER).inc();
-      }
-      txnHandler.markFailed(ci);
-    }
-  }
-
   private String getPoolName(CompactionInfo ci, Table t) throws Exception {
     Map<String, String> params = t.getParameters();
     String poolName = params == null ? null : params.get(Constants.HIVE_COMPACTOR_WORKER_POOL);
@@ -294,16 +262,7 @@ public class Initiator extends MetaStoreCompactorThread {
     return CompactorUtil.resolveDatabase(conf, ci.dbname);
   }
 
-  private ValidWriteIdList resolveValidWriteIds(Table t) throws NoSuchTxnException, MetaException {
-    ValidTxnList validTxnList = new ValidReadTxnList(conf.get(ValidTxnList.VALID_TXNS_KEY));
-    // The response will have one entry per table and hence we get only one ValidWriteIdList
-    String fullTableName = TxnUtils.getFullTableName(t.getDbName(), t.getTableName());
-    GetValidWriteIdsRequest rqst = new GetValidWriteIdsRequest(Collections.singletonList(fullTableName));
-    rqst.setValidTxnList(validTxnList.writeToString());
 
-    return TxnUtils.createValidCompactWriteIdList(
-        txnHandler.getValidWriteIds(rqst).getTblValidWriteIds().get(0));
-  }
 
   @VisibleForTesting
   protected String resolveUserToRunAs(Map<String, String> cache, Table t, Partition p)
@@ -394,164 +353,6 @@ public class Initiator extends MetaStoreCompactorThread {
     return false;
   }
 
-  private CompactionType checkForCompaction(final CompactionInfo ci,
-                                            final ValidWriteIdList writeIds,
-                                            final StorageDescriptor sd,
-                                            final Map<String, String> tblproperties,
-                                            final String runAs)
-      throws IOException, InterruptedException {
-    // If it's marked as too many aborted, we already know we need to compact
-    if (ci.tooManyAborts) {
-      LOG.debug("Found too many aborted transactions for " + ci.getFullPartitionName() + ", " +
-          "initiating major compaction");
-      return CompactionType.MAJOR;
-    }
-
-    if (ci.hasOldAbort) {
-      HiveConf.ConfVars oldAbortedTimeoutProp =
-          HiveConf.ConfVars.HIVE_COMPACTOR_ABORTEDTXN_TIME_THRESHOLD;
-      LOG.debug("Found an aborted transaction for " + ci.getFullPartitionName()
-          + " with age older than threshold " + oldAbortedTimeoutProp + ": " + conf
-          .getTimeVar(oldAbortedTimeoutProp, TimeUnit.HOURS) + " hours. "
-          + "Initiating minor compaction.");
-      return CompactionType.MINOR;
-    }
-    AcidDirectory acidDirectory = getAcidDirectory(sd, writeIds);
-    long baseSize = getBaseSize(acidDirectory);
-    FileSystem fs = acidDirectory.getFs();
-    Map<Path, Long> deltaSizes = new HashMap<>();
-    for (AcidUtils.ParsedDelta delta : acidDirectory.getCurrentDirectories()) {
-      deltaSizes.put(delta.getPath(), getDirSize(fs, delta));
-    }
-    long deltaSize = deltaSizes.values().stream().reduce(0L, Long::sum);
-    AcidMetricService.updateMetricsFromInitiator(ci.dbname, ci.tableName, ci.partName, conf, txnHandler,
-        baseSize, deltaSizes, acidDirectory.getObsolete());
-
-    if (runJobAsSelf(runAs)) {
-      return determineCompactionType(ci, acidDirectory, tblproperties, baseSize, deltaSize);
-    } else {
-      LOG.info("Going to initiate as user " + runAs + " for " + ci.getFullPartitionName());
-      UserGroupInformation ugi = UserGroupInformation.createProxyUser(runAs,
-        UserGroupInformation.getLoginUser());
-      CompactionType compactionType;
-      try {
-        compactionType = ugi.doAs(
-            (PrivilegedExceptionAction<CompactionType>) () -> determineCompactionType(ci, acidDirectory, tblproperties, baseSize, deltaSize));
-      } finally {
-        try {
-          FileSystem.closeAllForUGI(ugi);
-        } catch (IOException exception) {
-          LOG.error("Could not clean up file-system handles for UGI: " + ugi + " for " +
-              ci.getFullPartitionName(), exception);
-        }
-      }
-      return compactionType;
-    }
-  }
-
-  private AcidDirectory getAcidDirectory(StorageDescriptor sd, ValidWriteIdList writeIds) throws IOException {
-    Path location = new Path(sd.getLocation());
-    FileSystem fs = location.getFileSystem(conf);
-    return AcidUtils.getAcidState(fs, location, conf, writeIds, Ref.from(false), false);
-  }
-
-  private CompactionType determineCompactionType(CompactionInfo ci, AcidDirectory dir, Map<String,
-      String> tblproperties, long baseSize, long deltaSize) {
-    boolean noBase = false;
-    List<AcidUtils.ParsedDelta> deltas = dir.getCurrentDirectories();
-    if (baseSize == 0 && deltaSize > 0) {
-      noBase = true;
-    } else {
-      String deltaPctProp = tblproperties.get(COMPACTORTHRESHOLD_PREFIX +
-          HiveConf.ConfVars.HIVE_COMPACTOR_DELTA_PCT_THRESHOLD);
-      float deltaPctThreshold = deltaPctProp == null ?
-          HiveConf.getFloatVar(conf, HiveConf.ConfVars.HIVE_COMPACTOR_DELTA_PCT_THRESHOLD) :
-          Float.parseFloat(deltaPctProp);
-      boolean bigEnough =   (float)deltaSize/(float)baseSize > deltaPctThreshold;
-      boolean multiBase = dir.getObsolete().stream()
-              .anyMatch(path -> path.getName().startsWith(AcidUtils.BASE_PREFIX));
-
-      boolean initiateMajor =  bigEnough || (deltaSize == 0  && multiBase);
-      if (LOG.isDebugEnabled()) {
-        StringBuilder msg = new StringBuilder("delta size: ");
-        msg.append(deltaSize);
-        msg.append(" base size: ");
-        msg.append(baseSize);
-        msg.append(" multiBase ");
-        msg.append(multiBase);
-        msg.append(" deltaSize ");
-        msg.append(deltaSize);
-        msg.append(" threshold: ");
-        msg.append(deltaPctThreshold);
-        msg.append(" delta/base ratio > ").append(HiveConf.ConfVars.HIVE_COMPACTOR_DELTA_PCT_THRESHOLD.varname)
-            .append(": ");
-        msg.append(bigEnough);
-        msg.append(".");
-        if (!initiateMajor) {
-          msg.append("not");
-        }
-        msg.append(" initiating major compaction.");
-        LOG.debug(msg.toString());
-      }
-      if (initiateMajor) return CompactionType.MAJOR;
-    }
-
-    String deltaNumProp = tblproperties.get(COMPACTORTHRESHOLD_PREFIX +
-        HiveConf.ConfVars.HIVE_COMPACTOR_DELTA_NUM_THRESHOLD);
-    int deltaNumThreshold = deltaNumProp == null ?
-        HiveConf.getIntVar(conf, HiveConf.ConfVars.HIVE_COMPACTOR_DELTA_NUM_THRESHOLD) :
-        Integer.parseInt(deltaNumProp);
-    boolean enough = deltas.size() > deltaNumThreshold;
-    if (!enough) {
-      LOG.debug("Not enough deltas to initiate compaction for table=" + ci.tableName + "partition=" + ci.partName
-          + ". Found: " + deltas.size() + " deltas, threshold is " + deltaNumThreshold);
-      return null;
-    }
-    // If there's no base file, do a major compaction
-    LOG.debug("Found " + deltas.size() + " delta files, and " + (noBase ? "no" : "has") + " base," +
-        "requesting " + (noBase ? "major" : "minor") + " compaction");
-
-    return noBase || !isMinorCompactionSupported(tblproperties, dir) ?
-            CompactionType.MAJOR : CompactionType.MINOR;
-  }
-
-  private long getBaseSize(AcidDirectory dir) throws IOException {
-    long baseSize = 0;
-    if (dir.getBase() != null) {
-      baseSize = getDirSize(dir.getFs(), dir.getBase());
-    } else {
-      for (HdfsFileStatusWithId origStat : dir.getOriginalFiles()) {
-        baseSize += origStat.getFileStatus().getLen();
-      }
-    }
-    return baseSize;
-  }
-
-  private long getDirSize(FileSystem fs, ParsedDirectory dir) throws IOException {
-    return dir.getFiles(fs, Ref.from(false)).stream()
-        .map(HdfsFileStatusWithId::getFileStatus)
-        .mapToLong(FileStatus::getLen)
-        .sum();
-  }
-
-  private void requestCompaction(CompactionInfo ci, String runAs) throws MetaException {
-    CompactionRequest rqst = new CompactionRequest(ci.dbname, ci.tableName, ci.type);
-    if (ci.partName != null) rqst.setPartitionname(ci.partName);
-    rqst.setRunas(runAs);
-    if (StringUtils.isEmpty(ci.initiatorId)) {
-      rqst.setInitiatorId(getInitiatorId(Thread.currentThread().getId()));
-    } else {
-      rqst.setInitiatorId(ci.initiatorId);
-    }
-    rqst.setInitiatorVersion(this.runtimeVersion);
-    rqst.setPoolName(ci.poolName);
-    LOG.info("Requesting compaction: " + rqst);
-    CompactionResponse resp = txnHandler.compact(rqst);
-    if(resp.isAccepted()) {
-      ci.id = resp.getId();
-    }
-  }
-
   // Check if it's a dynamic partitioning case. If so, do not initiate compaction for streaming ingest, only for aborts.
   private static boolean isDynPartIngest(Table t, CompactionInfo ci){
     if (t.getPartitionKeys() != null && t.getPartitionKeys().size() > 0 &&
@@ -639,13 +440,6 @@ public class Initiator extends MetaStoreCompactorThread {
       return false;
     }
     return true;
-  }
-
-  private String getInitiatorId(long threadId) {
-    StringBuilder name = new StringBuilder(this.hostName);
-    name.append("-");
-    name.append(threadId);
-    return name.toString();
   }
 
   private static class InitiatorCycleUpdater implements Runnable {
